Various time related bugfixes as well as state serialization fix.
The switch from clock count to seconds was regressive in that it did not
make changes on "frames (samples) to count" calculation (that should have
been changed to "frames (samples) to seconds" calculation).

mixer state management had regressed on whether it shifts by 5 bits (for
noise_freq) or not.

Lastly we skip a lot of samples for calculating volume. We don't need
too sample-precise envelope.